+++ packages/eas-cli/src/credentials/android/actions/CreateGoogleServiceAccountKey.ts
@@ -1,5 +1,8 @@
+import chalk from 'chalk';
+import fs from 'fs-extra';
+
 import { GoogleServiceAccountKeyFragment } from '../../../graphql/generated';
-import Log from '../../../log';
+import Log, { learnMore } from '../../../log';
 import { promptAsync } from '../../../prompts';
 import { Account } from '../../../user/Account';
 import { CredentialsContext } from '../../context';
@@ -41,14 +44,33 @@ export class CreateGoogleServiceAccountKey {
       return detectedPath;
     }
 
-    const { keyJsonPath } = await promptAsync([
-      {
-        type: 'text',
-        name: 'keyJsonPath',
-        message: 'Path to Google Service Account Key JSON file:',
-        validate: (value: string) => value.length > 0 || "Path can't be empty",
+    Log.log(
+      `${chalk.bold(
+        'A Google Service Account JSON key is required to upload your app to Google Play Store'
+      )}.\n` +
+        `If you're not sure what this is or how to create one, ${learnMore(
+          'https://expo.fyi/creating-google-service-account',
+          { learnMoreMessage: 'learn more' }
+        )}`
+    );
+    const { filePath } = await promptAsync({
+      name: 'filePath',
+      message: 'Path to Google Service Account file:',
+      initial: 'api-0000000000000000000-111111-aaaaaabbbbbb.json',
+      type: 'text',
+      // eslint-disable-next-line async-protect/async-suffix
+      validate: async (filePath: string) => {
+        try {
+          const stats = await fs.stat(filePath);
+          if (stats.isFile()) {
+            return true;
+          }
+          return 'Input is not a file.';
+        } catch {
+          return 'File does not exist.';
+        }
       },
-    ]);
-    return keyJsonPath;
+    });
+    return filePath;
   }
 }